+++ CHANGELOG.md
@@ -26,6 +26,7 @@ See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog
 ### 🐛 Bug Fixes
 
 - Restore numeric keyboard on amount fields after Material UI update (#242)
+- Fix outlined TextField label overlapping the border on iOS Safari (#243)
 
 ## [2.2.0] - 2026-08-02